+++ server/DataBase.js
@@ -32,20 +32,7 @@ class DataBase{
 }
 
 
-class CookieDataBase{
-    constructor(){
-        this.cookieSaver = new Map();
-    }  
 
-    addCookie(id){
-        this.cookieSaver.set(id,1)
-    }
-
-    checkCookie(id){
-        return this.cookieSaver.has(id);
-    }
-
-}
 
 
 module.exports = DataBase , CookieDataBase
